requirements.txt
Pillow
django
djangorestframework
django-cors-headers
django-webpack-loader